suppress code analysis warnings

+++ ttcr/Grid2Drc.h
@@ -59,8 +59,8 @@ namespace ttcr {
         dx(ddx), dz(ddz), xmin(minx), zmin(minz),
         xmax(minx+nx*ddx), zmax(minz+nz*ddz),
         ncx(nx), ncz(nz),
-        nodes(std::vector<NODE>( (ncx+1) * (ncz+1), NODE(nt) )),
-        cells(ncx*ncz)
+        nodes(std::vector<NODE>( static_cast<size_t>(ncx+1) * (ncz+1), NODE(nt) )),
+        cells(static_cast<size_t>(ncx)*ncz)
         {
         }
 
@@ -68,8 +68,8 @@ namespace ttcr {
         }
 
         void getSlowness(std::vector<T1>& slowness) const {
-            if (slowness.size() != ncx*ncz) {
-                slowness.resize(ncx*ncz);
+            if (slowness.size() != static_cast<size_t>(ncx)*ncz) {
+                slowness.resize(static_cast<size_t>(ncx)*ncz);
             }
             for (size_t n=0; n<slowness.size(); ++n) {
                 slowness[n] = cells.getSlowness(n);
@@ -135,15 +135,15 @@ namespace ttcr {
 
         size_t getNumberOfNodes(const bool primary=false) const {
             if ( primary ) {
-                return (ncx+1) * (ncz+1);
+                return static_cast<size_t>(ncx+1) * (ncz+1);
             } else {
                 return nodes.size();
             }
         }
-        size_t getNumberOfCells() const { return ncx*ncz; }
+        size_t getNumberOfCells() const { return static_cast<size_t>(ncx)*ncz; }
 
         void getTT(std::vector<T1>& tt, const size_t threadNo=0) const {
-            size_t nPrimary = (ncx+1) * (ncz+1);
+            size_t nPrimary = static_cast<size_t>(ncx+1) * (ncz+1);
             tt.resize(nPrimary);
             size_t n = 0;
             for ( size_t nn=0; nn<nodes.size(); ++nn ) {
@@ -189,7 +189,7 @@ namespace ttcr {
         const T2 getNcz() const { return ncz; }
 
         void dump_secondary(std::ofstream& os) const {
-            size_t nPrimary = (ncx+1) * (ncz+1);
+            size_t nPrimary = static_cast<size_t>(ncx+1) * (ncz+1);
             for ( size_t n=nPrimary; n<nodes.size(); ++n ) {
                 os << nodes[n].getX() << ' ' << nodes[n].getZ() << '\n';
             }
