+++ ttcr/Grid2Drcsp.h
@@ -111,7 +111,7 @@ namespace ttcr {
         const T2 getNsnz() const { return nsnz; }
 
         void getTT(std::vector<T1>& tt, const size_t threadNo=0) const final {
-            size_t nPrimary = (this->ncx+1) * (this->ncz+1);
+            size_t nPrimary = static_cast<size_t>(this->ncx+1) * (this->ncz+1);
             tt.resize(nPrimary);
             for ( size_t n=0, n2=0; n<this->nodes.size(); ++n ) {
                 if (this->nodes[n].isPrimary()) {
