+++ ttcr/Grid2Drndsp.h
@@ -48,7 +48,7 @@ namespace ttcr {
             this->template buildGridNeighbors<Node2Dn<T1,T2>>(this->nodes);
             nPermanent = static_cast<T2>(this->nodes.size());
             for ( size_t n=0; n<nt; ++n ) {
-                tempNeighbors[n].resize(this->ncx * this->ncz);
+                tempNeighbors[n].resize(static_cast<size_t>(this->ncx) * this->ncz);
             }
             if (useEdgeLength) {
                 T1 dx = 0.5 * (ddx+ddz);
